projects/social-media-backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
graphene==3.3
graphene-django==3.1.1
django-cors-headers==3.14.0
django-filter==23.1
python-decouple==3.8
celery==5.3.0
redis==4.5.5
pillow==9.5.0
pytest==7.3.1
pytest-django==4.5.2
factory-boy==3.2.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
django-redis==5.2.0
django-ratelimit==4.1.0
